fix: seprate HeartOutline icons
